pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "googletrans>=4.0.2",
    "openai>=1.76.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "streamlit>=1.44.1",
]
